refactor: move Printer and Render directly into lib.rs
